detector_v2/app/scanner.py
import os
import aiohttp
from PIL import Image
import io

REBRICKABLE_API_KEY = os.getenv('REBRICKABLE_API_KEY')
REBRICKABLE_BASE_URL = "https://rebrickable.com/api/v3"

async def scan_minifigure(image_data):
    # Convert bytes to image for processing
    image = Image.open(io.BytesIO(image_data))
    
    # Here you would add your image processing logic
    # For now, we'll just query Rebrickable API with a test ID
    
    async with aiohttp.ClientSession() as session:
        url = f"{REBRICKABLE_BASE_URL}/lego/minifigs/fig-000001/"
        headers = {"Authorization": f"key {REBRICKABLE_API_KEY}"}
        
        async with session.get(url, headers=headers) as response:
            if response.status == 200:
                data = await response.json()
                return {
                    "is_official": True,
                    "name": data.get("name"),
                    "rebrickable_id": data.get("set_num")
                }
            
    return {
        "is_official": False,
        "message": "No matching minifigure found"
    }
