detector_v2/requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
google-cloud-storage
python-multipart
pillow
aiohttp
